Move toaq.peg to toaq.peggy; make toaq.peg a "pretty" form of the grammar.
The grammar is starting to get messy as actions are added.
It's nice to have a simplified, readable version of the grammar.
The simplified version has the .peg extension, since it's basically normal PEG syntax.
The version of the grammar containing all of the actions gets the .peggy extension,
because it's the input readable by the parser generator, Peggy.