Fix statement and predicate forethought CoPs so that they are no longer snarfed by sentences.
The sentence and statement forethough CoPs eagrly matched a forethought connective, and never gave a chance for statement and predicate CoPs to match respectively. The effect is that would-be statement and predicate forethought CoPs always parsed as sentence CoPs. Now they do not. The highest-precedence element parsable in the left-branch of the CoP is what is connected.