ast: disallow prenext statements on the right-hand-side of a statement afterthought connective.
Allowing RHS prenex statements created unintuitive parses for content clauses like lî shỉ na ra lî dẻ na bi hỏq.
The parse of this used to connect **shi?** as a predication statement with **li^ de? na bi hoq?** as prenex statement, and the whole thing became a fragment when wrappend in the first **li^**. However, the expected parse (according to a bunch of us) would have this as a statement with two **ru**-connected content clauses in the prenex.

Fixes #18